feat(wasm): make hf-xet compile for wasm32-unknown-unknown (#841)
Closes #840

## Summary

Makes `hf-xet` (`xet_pkg`) and its streaming download + upload data-prep
paths compile and **run** on `wasm32-unknown-unknown`, and reworks
`wasm/hf_xet_wasm/` into a single example/smoke-test JS-binding crate
(`cdylib`) that exposes both flows from one wasm module. (The earlier
split into separate download/upload crates was reverted — one module is
enough for the smokes and lets the example pages share it.)
`wasm/hf_xet_thin_wasm/` is unchanged in scope but follows the new
shared wasm-bindgen pinning (`=0.2.121`).

## Approach

1. **Compile compatibility** — cfg-gated `tokio` dep split; native-only
gating of `_blocking`/path-based methods, `xet_pkg::legacy`, and native
runtime pieces; `xet_client` cache gating.
2. **Streaming runtime compatibility** — make the streaming
download/upload paths `!Send`-tolerant so they actually execute on wasm
(where reqwest's futures are `!Send`), via conditional `?Send`
async-traits and `use tokio_with_wasm::alias as tokio` at every spawn
site.

A later simplification pass deduplicates the wasm/native split where the
bodies were identical: `TaskRuntime::bridge_async{,_finalizing}` are
written once via a `MaybeSend` shim, `XetSessionBuilder::build` /
`TranslatorConfig::new` are single cfg-bound fns, the wasm shard dedup
cache is a plain bounded FIFO, and the ci-smoke suite runs from one
shared harness.

## Key changes

- **`tokio_with_wasm`** added to workspace deps and the wasm-target deps
of `xet_pkg`/`xet_data`/`xet_runtime`/`xet_client`. Native uses real
`tokio`; wasm aliases to `tokio_with_wasm` (spawns via
`wasm_bindgen_futures::spawn_local`, so spawned futures needn't be
`Send`).
- **Conditional `?Send` async-traits** for the streaming
providers/writers (`URLProvider`, `DataWriter`, `XorbURLProvider`,
`Sequential/UnorderedWriter`, `UploadSessionDataManager`). Trait bounds
on `dyn T` stay `Send + Sync`; only the futures drop `Send` on wasm.
- **Native-only gating** of all `_blocking`/filesystem methods,
`XetSession::sigint_abort`, `XetSessionBuilder::with_tokio_handle`, the
`xorb_block.rs` chunk-cache spawn, and the entire `xet_pkg::legacy`
module — surfaced as compile errors rather than runtime panics.
- **`xet_runtime`**: the wasm `XetRuntime` is a stub (no tokio runtime;
bridges `.await` directly via `wasm_bindgen_futures`); `spawn_blocking`
maps to `tokio_with_wasm`; native-only accessors and the `SystemMonitor`
are gated off on wasm.
- **`xet_data`**: `processing` + `file_reconstruction` are now available
on wasm; new `shard_interface/wasm.rs` provides an in-memory
`SessionShardInterface` with a bounded FIFO dedup cache (no disk
staging, no resume).
- **`wasm/hf_xet_wasm/`** (unified `cdylib`): `#[wasm_bindgen]` wrapper
over `xet::xet_session::XetSession` exposing both upload and download
from one module. Auth-free `XetSession`; auth lives on the per-commit /
per-group builder. Ships `build_wasm.sh` +
`examples/{download,upload}.html`. **Not a published browser SDK** —
real consumers should depend on `hf-xet` directly with their own
`#[wasm_bindgen]` glue, or use a downstream SDK such as `hf-hub`.
- **CI** (`.github/actions/build-wasm` + `ci.yml`): builds both wasm
crates, runs the `xet_pkg/build_wasm.sh` compile gate (`cargo +nightly
check --target wasm32-unknown-unknown -p hf-xet -Z build-std`), checks
the wasm `Cargo.lock`s are fresh, then runs the `wasm/ci-smoke/`
headless-Chromium Playwright suite via `node run.mjs all`.

## WASM API surface (downloads)

```
XetSessionBuilder::build() → XetSession
XetSession::new_download_stream_group() → XetDownloadStreamGroupBuilder
  .with_endpoint(...)
  .with_token_info(token, expiry)
  .with_token_refresh_url(...)
  .build().await → XetDownloadStreamGroup
XetDownloadStreamGroup::download_stream(file_info, range).await → XetDownloadStream
XetDownloadStream::next().await → Option<Bytes>
XetDownloadStream::cancel()
XetSession::abort() / id() / config()
```

## WASM API surface (uploads)

```
XetSessionBuilder::build() → XetSession
XetSession::new_upload_commit() → XetUploadCommitBuilder
  .with_endpoint(...)
  .with_token_info(token, expiry)
  .build().await → XetUploadCommit
XetUploadCommit::upload_bytes(bytes, sha256, name).await → XetFileUpload
XetUploadCommit::upload_stream(name, sha256).await → XetStreamUpload
XetStreamUpload::write(chunk).await / finish().await / abort()
XetUploadCommit::commit().await → XetCommitReport
XetUploadCommit::abort()
```

The unified `hf_xet_wasm` crate exposes the JS classes `XetSession`,
`XetDownloadStreamGroup`, `XetDownloadStream`, `XetUploadCommit`,
`XetStreamUpload`, and `XetFileUpload` from a single wasm module.

## Testing

- **Native**: `hf-xet` (95 unit + 6 doctest) and `xet-data` (283 unit)
tests pass.
- **WASM compile + build**: `xet_pkg/build_wasm.sh` (the wasm compile
gate) and both crates' `build_wasm.sh` scripts produce
`pkg/{js,d.ts,wasm}` (CI).
- **WASM browser smokes** (CI, headless Chromium + Playwright): **18
scenarios** in `wasm/ci-smoke/`, run in a single `node run.mjs all` step
against prod Hub + CAS (`HF_SMOKE_TEST_TOKEN`). Network scenarios get
one retry; any failure fails the job. Coverage: local input validation;
single / multi / range / cancel / error downloads (streaming +
`downloadToBytes`); single / multi / mixed / tiny / streaming uploads;
multi-commit; `sha256-policy`; upload lifecycle/misuse; session-level
dedup; and HMAC-keyed global dedup (re-upload fully dedups —
`xorb_bytes_uploaded === 0`, no new xorb).
- **Manual**: serve `wasm/hf_xet_wasm/examples/{download,upload}.html`
with a static server (the smoke `server.mjs` sets the required COOP/COEP
headers).


<!-- CURSOR_SUMMARY -->
---

> [!NOTE]
> **High Risk**
> Large cross-cutting changes to upload/download, dedup, HTTP, and async
runtime behavior on wasm, plus CI that hits production Hub/CAS with
write-capable tokens.
> 
> **Overview**
> Adds **`wasm32-unknown-unknown`** support for `hf-xet` and its
transitive crates so browser consumers can depend on the library
directly, with a **wasm-only public API subset** (async-only, no
filesystem/`legacy`, stub `XetRuntime`).
> 
> **Runtime / async:** Introduces **`tokio_with_wasm`** and
workspace-pinned **`wasm-bindgen` 0.2.121**; wasm code uses
`tokio_with_wasm::alias as tokio`, conditional **`async_trait(?Send)`**,
and **`web_time`** instead of std/tokio instants. Native blocking APIs,
path downloads, disk caches, and **`xet_pkg::legacy`** are
**`#[cfg(not(target_family = "wasm"))]`**. **`xet_data`** gains an
in-memory **`shard_interface/wasm`** path;
**`RemoteClient::upload_xorb`** uses non-streaming bodies on wasm with
bulk progress reporting.
> 
> **`wasm/hf_xet_wasm`:** Reworked into a thin **`cdylib`** over
`hf-xet` (unified upload/download JS smoke wrapper, not a published
SDK). **`wasm/ci-smoke/`** adds **18 Playwright scenarios** against prod
Hub/CAS, run from CI via **`node run.mjs all`** after wasm builds and
**`xet_pkg/build_wasm.sh`** compile gate.
> 
> **CI / tooling:** **`build-wasm`** caches wasm-bindgen/wasm-pack
binaries, bumps tool versions, and runs the **`hf-xet`** wasm check;
docs and **`api_changes/update_260515_wasm_target_support.md`** document
wasm differences and maintenance patterns.
> 
> <sup>Reviewed by [Cursor Bugbot](https://cursor.com/bugbot) for commit
d549241f78d91d75dfb8eae4a54bfa86f203aca2. Bugbot is set up for automated
code reviews on this repo. Configure
[here](https://www.cursor.com/dashboard/bugbot).</sup>
<!-- /CURSOR_SUMMARY -->

---------

+++ xet_data/src/deduplication/file_deduplication.rs
@@ -96,7 +96,7 @@ impl<DataInterfaceType: DeduplicationDataInterface> FileDeduper<DataInterfaceTyp
             .config
             .xorb
             .simulation_max_bytes
-            .map(|bs| (bs.as_u64() as usize).min(*MAX_XORB_BYTES))
+            .map(|bs| bs.as_u64().min(*MAX_XORB_BYTES as u64) as usize)
             .unwrap_or(*MAX_XORB_BYTES);
         #[cfg(not(feature = "simulation"))]
         let xorb_cut_bytes = *MAX_XORB_BYTES;
